Split the oracle's verdict layer onto a real seam (refs #423)
`capture_tableau_oracle.py` was 1161 lines against pylint's 1200-line ceiling, so
manifest and an exit code -- `log_progress`, `_render_statuses`, `_partition`,
`CaptureRun`, `write_manifest`, `_log_blocked_and_stale` -- into
`scripts/tableau_oracle_manifest.py`, the same way `tableau_view_types.py` and
`tableau_payload_facts.py` were split out before it. 1161 -> 875 lines.

The seam is real, not arithmetic: nothing in the new module talks to Tableau. It
wants two counters and a redactor off the session, so it takes the session
duck-typed and the pair stays acyclic -- the arrangement `tableau_render_capability`
already uses. `SVG_MIN_API_VERSION` / `SVG_VERSION_MARKER` travel with it because
both exist to CLASSIFY a failure and name the knob that fixes it, which is verdict
vocabulary rather than transport.

No behaviour change: the moved code is byte-identical apart from the `CaptureRun.session`
annotation (`TableauSession` -> `Any`, which the docstring now states as the duck type,
because the real annotation would reinstate the cycle).

The new module joins `tests/test_diagnostic_redaction.py`'s MODULES, because it now holds
THE SINK -- `scrub_tree` immediately before `json.dumps` -- plus every console line that
quotes a response-derived view name. Leaving it out would have moved the most sensitive
code in the capture OUT of that gate while every parameterised check kept reporting green.

That gate is deliberately WIDENED, not merely re-pointed. Seeding `write_manifest`'s
`records` exposed a pre-existing blind spot: on master the list was never tainted, because
`main` binds it with `records, started = [], ...` and then grows it with `.append()`, and
`_bind` fires on assignment rather than on a container-store call -- so a list full of
response records crossed into `write_manifest` analysed as clean, and only `capability_report`
made the manifest tainted at all. Twenty-odd expressions now require a certification; all are
counts, status literals from a closed set, or the aggregate `scrub_tree` walks whole.
`tableau_view_types.census(records)` became a declared cross-module seed for the same reason.

Gates, by exit code: ruff format/check 0, `pylint scripts` 0 (10.00/10), and the full
`pytest tests/` 3613 passed / 6 skipped (excluding the known `test_upstream_repro_pins.py`
engine-drift tripwires).
